% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_da.R
\name{ap_plot_concordance}
\alias{ap_plot_concordance}
\title{Concordance between methods}
\usage{
ap_plot_concordance(concordance, max_features = 40L)
}
\arguments{
\item{concordance}{An \code{ap_da_concordance} result from \code{\link[=ap_da_concordance]{ap_da_concordance()}}.}

\item{max_features}{Cap on features drawn. Default \code{40}.}
}
\value{
A ggplot object.
}
\description{
Which methods called which features, as a presence grid ordered by how many
methods agreed.
}
